requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
Django==4.1.1
django-apscheduler==0.6.2
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
numpy==1.23.3
pandas==1.4.4
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
tzlocal==4.2
